% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMeta.plot.R
\name{GMeta.plot}
\alias{GMeta.plot}
\title{Generalized Meta-analysis(forest plot)}
\usage{
GMeta.plot(x, study_info_plot)
}
\arguments{
\item{x}{an object of class "GMeta"}

\item{study_info_plot}{a list of lists containing information about the studies(similar to the study_info argument used in GMeta function.)}
}
\description{
This function plots the confidence intervals with boxes as the study specific estimates and diamond as the GMeta estimate. For the current version, it assumes that the estimate of the variance-covariance matrix in each of the studies is provided.
It is demonstrated using a different dataset, "study_info_plot", which meets the assumption.
}
\examples{
# This example shows how to obtain the forest plot of GMeta object.
data(study_info_plot)
data(reference_data)
model <- "logistic"
result_diff <- GMeta(study_info_plot, reference_data, model, variable_intercepts = TRUE)
GMeta.plot(result_diff, study_info_plot)
}
