% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMeta.R
\name{GMeta}
\alias{GMeta}
\title{Implementing Generalized Meta-analysis}
\usage{
GMeta(study_info, ref_dat, model, variable_intercepts = FALSE,
  control = list(...))
}
\arguments{
\item{study_info}{a list of lists containing information about the studies; the main list contains a list for each study, which must have the fields:
\itemize{
\item{"Coeff": a named numeric vector containing the estimates of regression parameters (including intercept) where the names identify the covariates. For example, names(study_info$Coeff) <- c("(Intercept)", "Age", "Height", "Weight").}
\item{"Covariance": a matrix containing an estimate of variance-covariance matrix of the regression parameters. This can be NULL if the "Sample_size" is provided.}
\item{"Sample_size": a numeric containing sample size of the study. This can be NULL if the "Covariance" is provided.}
}}

\item{ref_dat}{a data matrix containing all the distinct covariates across studies from a reference set of individuals. This is used for estimating joint distribution of the covariates. The data matrix must have the vector of ones as its first column. The column names of the data matrix should match the names of the covariates from the studies.}

\item{model}{a description of the type of regression model; this is a character string naming the regression model. The current version is for "logistic" and "linear".}

\item{variable_intercepts}{an optional logical (applicable only when the model is "logistic"); if TRUE, the intercepts of the true models for each of the studies are assumed to be different. Default is FALSE.}

\item{control}{an optional list containing the epsilon (numeric) and maxiter (numeric) needed for convergence of the algorithm. Default epsilon and maximum iterations are 1e-06 and 1000, respectively. For creating a control argument for GMeta, see \code{\link[GMeta]{GMeta.control}}.}
}
\value{
An object of class "GMeta" is a list containing GMeta estimate, its variance-covariance matrix and estimates the residual variance in the case of "linear" model .
\item{Est.coeff}{a numeric vector containing the estimated regression coefficients of the maximal model using optimal weighting matrix.}
\item{Est.var.cov}{a matrix containing estimate of variance-covariance matrix of the corresponding GMeta estimator.}
\item{Res.var}{a numeric containing the residual variance of the maximal model when it is linear. It is calculated from the formula : \eqn{1 - \hat{\beta}^Tvar(X)\hat{\beta}} which is derived by assuming the outcomes to have unit variance. \eqn{var(X)} is calculated from reference data.}
\item{iter}{a numeric containing the number of iterations used in the algorithm}
\cr Res.var is NA when the model is "logistic".
\cr The function \code{\link[GMeta]{GMeta.summary}} can be used to obtain a summary of the results obtained from GMeta.
}
\description{
Generalized Meta-analysis(GMeta) is an approach for combining information on multivariate regression parameters across multiple different studies which have different, but, possibly overlapping information on subsets of covariates.
GMeta implements the generalized meta-analysis using IRWLS algorithm.
}
\details{
Generalized Meta-analysis (GMeta) is a tool that allows researchers to quickly build models for multivariate meta-analysis in the presence of disparate covariate information across studies. It is implemented based on mainly two input arguments:
\itemize{
\item{Information on the model parameters from each of the studies.}
\item{Reference data for estimation of the joint distribution of all the distinct covariates across studies.}}
The software provides flexibility to the users to choose the intercepts to be different (when the model is logistic) across studies through the input argument, variable_intercepts.
It also allows estimation of the regression parameters, only from the sample sizes of the studies when it is difficult to obtain estimate of the variance-covariance matrices.
\cr \cr \bold{Note}: GMeta will not work if both the estimates of the covariance matrix and the sample size are NULL.

When the model is "linear", it is assumed that the outcome is standardized to have unit variance.
For more details on the IRWLS, see References.
}
\examples{
# This example shows the GMeta implementation on a simulated data set for logistic regression
data(reference_data)
head(reference_data)
data(study_info)
head(study_info)
model <- "logistic"
# When the true intercept parameters of the studies are different.
result_diff <- GMeta(study_info, reference_data, model, variable_intercepts = TRUE)
print(result_diff)
# When the true intercept parameters of the studies are same.
result_same <- GMeta(study_info, reference_data, model)
print(result_same)
}
\references{
Tang, R., Kundu, P. and Chatterjee, N. (2017) Generalized Meta-Analysis for Multivariate Regression Models Across Studies with Disparate Covariate Information. \href{https://arxiv.org/abs/1708.03818}{arXiv:1708.03818v1 [stat.ME]}.
}
\seealso{
\code{\link[GMeta]{GMeta.summary}}, \code{\link[GMeta]{study_info}}, \code{\link[GMeta]{reference_data}}
}
\author{
Runlong Tang, Prosenjit Kundu and Nilanjan Chatterjee.
}
\keyword{Analysis}
\keyword{Generalized}
\keyword{Meta}
