% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMeta.summary.R
\name{GMeta.summary}
\alias{GMeta.summary}
\title{Summarizing Generalized Meta Analysis}
\usage{
GMeta.summary(object, signi_digits = 3)
}
\arguments{
\item{object}{an object of class "GMeta"}

\item{signi_digits}{an optional numeric indicating the number of significant digits to be shown in the summary. Default is 3.}
}
\description{
This function prints the summary of GMeta results.
}
\examples{
# This example shows how to obtain the summary of GMeta object.
data(reference_data)
data(study_info)
model <- "logistic"
result_diff <- GMeta(study_info, reference_data, model, variable_intercepts = TRUE)
GMeta.summary(result_diff, signi_digits = 4)
result_same <- GMeta(study_info, reference_data, model)
GMeta.summary(result_same)
}
